+++ stream-compositions/services/legal-entity-composition-service/src/main/java/com/backbase/stream/compositions/legalentity/core/config/ProductCompositionClientConfiguration.java
@@ -1,0 +1,35 @@
+package com.backbase.stream.compositions.legalentity.core.config;
+
+import com.backbase.stream.clients.config.CompositeApiClientConfig;
+import com.backbase.stream.compositions.product.ApiClient;
+import com.backbase.stream.compositions.product.client.ProductCompositionApi;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import java.text.DateFormat;
+import org.springframework.boot.context.properties.ConfigurationProperties;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Primary;
+
+@Configuration
+@ConfigurationProperties("backbase.communication.services.stream.product.composition")
+public class ProductCompositionClientConfiguration extends CompositeApiClientConfig {
+
+    public static final String SERVICE_ID = "product-composition";
+
+    public ProductCompositionClientConfiguration() {
+        super(SERVICE_ID);
+    }
+
+    @Bean
+    public ApiClient productClient(ObjectMapper objectMapper, DateFormat dateFormat) {
+        return new ApiClient(getWebClient(), objectMapper, dateFormat)
+            .setBasePath(createBasePath());
+    }
+
+    @Bean
+    @Primary
+    public ProductCompositionApi productCompositionApi(ApiClient productClient) {
+        return new ProductCompositionApi(productClient);
+    }
+
+}