+++ stream-compositions/services/legal-entity-composition-service/src/test/java/com/backbase/stream/compositions/legalentity/core/config/LegalEntityConfigurationPropertiesTest.java
@@ -16,11 +16,9 @@ class LegalEntityConfigurationPropertiesTest {
     @Test
     void test() {
         LegalEntityConfigurationProperties properties = new LegalEntityConfigurationProperties();
-        properties.setIntegrationBaseUrl("https://legal-entity");
 
         Chains chains = new Chains();
         ProductComposition productComposition = new ProductComposition();
-        productComposition.setBaseUrl("https://product-composition");
         productComposition.setEnabled(Boolean.TRUE);
         productComposition.setAsync(Boolean.TRUE);
         chains.setProductComposition(productComposition);
@@ -30,22 +28,15 @@ class LegalEntityConfigurationPropertiesTest {
         events.setEnableFailed(Boolean.TRUE);
 
         Cursor cursor = new Cursor();
-        cursor.setBaseUrl("https://cursor");
         cursor.setEnabled(Boolean.TRUE);
 
         properties.setChains(chains);
         properties.setEvents(events);
         properties.setCursor(cursor);
 
-        assertTrue(properties.getIntegrationBaseUrl().contains("legal-entity"),
-                "Correct config spotted");
-        assertTrue(
-                properties.getChains().getProductComposition().getBaseUrl().contains("product-composition"),
-                "Correct config spotted");
         assertTrue(properties.getEvents().getEnableCompleted());
         assertTrue(properties.getEvents().getEnableFailed());
         assertTrue(properties.getCursor().getEnabled());
-        assertTrue(properties.getCursor().getBaseUrl().contains("cursor"), "Correct config spotted");
         assertTrue(properties.isCompletedEventEnabled());
         assertTrue(properties.isFailedEventEnabled());
         assertTrue(properties.isProductChainEnabled());