+++ stream-compositions/services/product-catalog-composition-service/src/main/java/com/backbase/stream/compositions/productcatalog/core/config/ProductCatalogConfiguration.java
@@ -1,11 +1,11 @@
 package com.backbase.stream.compositions.productcatalog.core.config;
 
-import com.backbase.buildingblocks.webclient.WebClientConstants;
+import com.backbase.stream.clients.config.CompositeApiClientConfig;
 import com.backbase.stream.compositions.integration.productcatalog.ApiClient;
 import com.backbase.stream.compositions.integration.productcatalog.api.ProductCatalogIntegrationApi;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import lombok.AllArgsConstructor;
-import org.springframework.beans.factory.annotation.Qualifier;
+import java.text.DateFormat;
+import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -13,16 +13,18 @@ import org.springframework.context.annotation.Primary;
 import org.springframework.security.config.annotation.web.reactive.EnableWebFluxSecurity;
 import org.springframework.security.config.web.server.ServerHttpSecurity;
 import org.springframework.security.web.server.SecurityWebFilterChain;
-import org.springframework.web.reactive.function.client.WebClient;
-
-import java.text.DateFormat;
 
 @Configuration
-@AllArgsConstructor
 @EnableWebFluxSecurity
 @EnableConfigurationProperties(ProductCatalogConfigurationProperties.class)
-public class ProductCatalogConfiguration {
-    private final ProductCatalogConfigurationProperties properties;
+@ConfigurationProperties("backbase.communication.services.stream.product-catalog.integration")
+public class ProductCatalogConfiguration extends CompositeApiClientConfig {
+
+    private static final String SERVICE_ID = "product-catalog-ingestion-integration";
+
+    public ProductCatalogConfiguration() {
+        super(SERVICE_ID);
+    }
 
     @Bean
     public SecurityWebFilterChain springSecurityFilterChain(ServerHttpSecurity http) {
@@ -36,13 +38,8 @@ public class ProductCatalogConfiguration {
     }
 
     @Bean
-    public ApiClient productCatalogClient(
-            @Qualifier(WebClientConstants.INTER_SERVICE_WEB_CLIENT_NAME) WebClient dbsWebClient,
-            ObjectMapper objectMapper,
-            DateFormat dateFormat) {
-        ApiClient apiClient = new ApiClient(dbsWebClient, objectMapper, dateFormat);
-        apiClient.setBasePath(properties.getProductCatalogIntegrationUrl());
-
-        return apiClient;
+    public ApiClient productCatalogClient(ObjectMapper objectMapper, DateFormat dateFormat) {
+        return new ApiClient(getWebClient(), objectMapper, dateFormat)
+            .setBasePath(createBasePath());
     }
 }