+++ stream-compositions/services/transaction-composition-service/src/test/java/com/backbase/stream/compositions/transaction/core/config/TransactionConfigurationPropertiesTest.java
@@ -1,13 +1,14 @@
 package com.backbase.stream.compositions.transaction.core.config;
 
-import com.backbase.stream.compositions.transaction.core.config.TransactionConfigurationProperties;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertFalse;
+import static org.junit.jupiter.api.Assertions.assertTrue;
+
 import com.backbase.stream.compositions.transaction.core.config.TransactionConfigurationProperties.Cursor;
 import com.backbase.stream.compositions.transaction.core.config.TransactionConfigurationProperties.Events;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.junit.jupiter.MockitoExtension;
-
-import static org.junit.jupiter.api.Assertions.*;
 
 @ExtendWith(MockitoExtension.class)
 class TransactionConfigurationPropertiesTest {
@@ -15,7 +16,6 @@ class TransactionConfigurationPropertiesTest {
     @Test
     void testConfig() {
         TransactionConfigurationProperties properties = new TransactionConfigurationProperties();
-        properties.setIntegrationBaseUrl("https://transaction");
         properties.setDefaultStartOffsetInDays(30);
 
         Events events = new Events();
@@ -23,21 +23,17 @@ class TransactionConfigurationPropertiesTest {
         events.setEnableFailed(Boolean.TRUE);
 
         Cursor cursor = new Cursor();
-        cursor.setBaseUrl("https://transaction-cursor");
         cursor.setTransactionIdsFilterEnabled(Boolean.FALSE);
         cursor.setEnabled(Boolean.TRUE);
 
         properties.setEvents(events);
         properties.setCursor(cursor);
 
-        assertTrue(properties.getIntegrationBaseUrl().contains("transaction"),
-                "Correct config spotted");
         assertEquals(30, properties.getDefaultStartOffsetInDays());
         assertTrue(properties.getEvents().getEnableCompleted());
         assertTrue(properties.getEvents().getEnableFailed());
         assertTrue(properties.getCursor().getEnabled());
         assertFalse(properties.getCursor().getTransactionIdsFilterEnabled());
-        assertTrue(properties.getCursor().getBaseUrl().contains("transaction-cursor"), "Correct config spotted");
         assertTrue(properties.isCursorEnabled());
         assertFalse(properties.isTransactionIdsFilterEnabled());
     }